+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EditorComponents no longer an interface.
 - EntityApplication no longer an interface.
+- EditorComponents bug fixed, factory method now public.
 
 ## 0.18.72
 ### is.codion.common.reactive